Duplicate settings/function to setup script
Having a settings.in.sh is a good in thing in development to have all
settings at one place but it's a burden on the live system. It's easier
to have only one script on the clients. Therefore the configuration and
printout function are duplicated for the setup_puppet-agent.sh script.